Add base64 method and name it TokenFormatter
